выдел интерфейсов
