+++ src/app/HomeWork/homeWork.component.ts
@@ -1,7 +1,7 @@
 import { NgFor, NgIf } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
 import { Component, inject, Injectable } from '@angular/core';
-import { User } from '../users-list/users-list.component';
+import { User } from '../user-interface';
 
 const newUsers = [
     {
