+++ src/app/todo-interface.ts
@@ -1,0 +1,6 @@
+export interface Todo {
+    userId: number;
+    id: number;
+    title: string;
+    completed: boolean
+}
